+++ src/App/Fossa/FossaAPIV1.hs
@@ -24,10 +24,16 @@ module App.Fossa.FossaAPIV1 (
   getSignedURL,
   archiveUpload,
   archiveBuildUpload,
+  assertUserDefinedBinaries,
+  assertRevisionBinaries,
+  resolveUserDefinedBinary,
+  resolveProjectDependencies,
 ) where
 
 import App.Fossa.Container (ContainerScan (..))
 import App.Fossa.Report.Attribution qualified as Attr
+import App.Fossa.VSI.IAT.Types qualified as IAT
+import App.Fossa.VSI.Types qualified as VSI
 import App.Types
 import App.Version (versionNumber)
 import Control.Carrier.Empty.Maybe (Empty, EmptyC, runEmpty)
@@ -452,3 +458,85 @@ uploadContributors apiOpts locator contributors = fossaReq $ do
 
   _ <- req POST (contributorsEndpoint baseUrl) (ReqBodyJson contributors) ignoreResponse opts
   pure ()
+
+----------
+
+-- | Core expects an object for each fingerprint.
+-- This type allows us to convert a fingerprint to the required object.
+newtype FingerprintSet = FingerprintSet
+  { fingerprintRaw :: IAT.Fingerprint
+  }
+
+instance ToJSON FingerprintSet where
+  toJSON FingerprintSet{..} =
+    object
+      [ "fingerprint_sha_256" .= fingerprintRaw
+      ]
+
+-- | The set of necessary data to describe a user defined binary assertion to the FOSSA API.
+data UserDefinedAssertionBody = UserDefinedAssertionBody
+  { bodyName :: Text
+  , bodyVersion :: Text
+  , bodyLicense :: Text
+  , bodyDescription :: Maybe Text
+  , bodyUrl :: Maybe Text
+  , bodyFingerprints :: [FingerprintSet]
+  }
+
+instance ToJSON UserDefinedAssertionBody where
+  toJSON UserDefinedAssertionBody{..} =
+    object
+      [ "name" .= bodyName
+      , "version" .= bodyVersion
+      , "license" .= bodyLicense
+      , "description" .= bodyDescription
+      , "url" .= bodyUrl
+      , "fingerprints" .= bodyFingerprints
+      ]
+
+assertUserDefinedBinariesEndpoint :: Url scheme -> Url scheme
+assertUserDefinedBinariesEndpoint baseurl = baseurl /: "api" /: "iat" /: "binary"
+
+assertUserDefinedBinaries :: (Has (Lift IO) sig m, Has Diagnostics sig m) => ApiOpts -> IAT.UserDefinedAssertionMeta -> [IAT.Fingerprint] -> m ()
+assertUserDefinedBinaries apiOpts IAT.UserDefinedAssertionMeta{..} fingerprints = fossaReq $ do
+  (baseUrl, baseOpts) <- useApiOpts apiOpts
+  let body = UserDefinedAssertionBody assertedName assertedVersion assertedLicense assertedDescription assertedUrl (FingerprintSet <$> fingerprints)
+  _ <- req POST (assertUserDefinedBinariesEndpoint baseUrl) (ReqBodyJson body) ignoreResponse baseOpts
+  pure ()
+
+assertRevisionBinariesEndpoint :: Url scheme -> Locator -> Url scheme
+assertRevisionBinariesEndpoint baseurl locator = baseurl /: "api" /: "iat" /: "binary" /: renderLocator locator
+
+assertRevisionBinaries :: (Has (Lift IO) sig m, Has Diagnostics sig m) => ApiOpts -> Locator -> [IAT.Fingerprint] -> m ()
+assertRevisionBinaries apiOpts locator fingerprints = fossaReq $ do
+  (baseUrl, baseOpts) <- useApiOpts apiOpts
+  let body = FingerprintSet <$> fingerprints
+  _ <- req POST (assertRevisionBinariesEndpoint baseUrl locator) (ReqBodyJson body) ignoreResponse baseOpts
+  pure ()
+
+resolveUserDefinedBinaryEndpoint :: Url scheme -> IAT.UserDep -> Url scheme
+resolveUserDefinedBinaryEndpoint baseurl dep = baseurl /: "api" /: "iat" /: "resolve" /: "user-defined" /: IAT.renderUserDep dep
+
+resolveUserDefinedBinary :: (Has (Lift IO) sig m, Has Diagnostics sig m) => ApiOpts -> IAT.UserDep -> m IAT.UserDefinedAssertionMeta
+resolveUserDefinedBinary apiOpts dep = fossaReq $ do
+  (baseUrl, baseOpts) <- useApiOpts apiOpts
+  responseBody <$> req GET (resolveUserDefinedBinaryEndpoint baseUrl dep) NoReqBody jsonResponse baseOpts
+
+-- | The revision dependencies endpoint contains a lot of information we don't need. This intermediate type allows us to throw it away.
+newtype ResolvedDependency = ResolvedDependency {unwrapResolvedDependency :: VSI.Locator}
+
+instance FromJSON ResolvedDependency where
+  parseJSON = withObject "ResolvedProjectDependencies" $ \obj -> do
+    ResolvedDependency <$> obj .: "loc"
+
+resolveProjectDependenciesEndpoint :: Url scheme -> VSI.Locator -> Url scheme
+resolveProjectDependenciesEndpoint baseurl locator = baseurl /: "api" /: "revisions" /: VSI.renderLocator locator /: "dependencies"
+
+resolveProjectDependencies :: (Has (Lift IO) sig m, Has Diagnostics sig m) => ApiOpts -> VSI.Locator -> m [VSI.Locator]
+resolveProjectDependencies apiOpts locator = fossaReq $ do
+  (baseUrl, baseOpts) <- useApiOpts apiOpts
+
+  let opts = baseOpts <> "include_ignored" =: False
+  (dependencies :: [ResolvedDependency]) <- responseBody <$> req GET (resolveProjectDependenciesEndpoint baseUrl locator) NoReqBody jsonResponse opts
+
+  pure $ map unwrapResolvedDependency dependencies